Move resetZoom method from backend to Plot
