+++ silx/gui/plot/backends/BackendBase.py
@@ -340,23 +340,6 @@ class BackendBase(object):
 
     # Graph limits
 
-    def resetZoom(self, dataMargins):
-        """Reset the displayed area of the plot.
-
-        Autoscale any axis that is in autoscale mode.
-        Keep current limits on axes not in autoscale mode
-
-        Extra margins can be added around the data inside the plot area.
-        Margins are given as one ratio of the data range per limit of the
-        data (xMin, xMax, yMin and yMax limits).
-        For log scale, extra margins are applied in log10 of the data.
-
-        :param dataMargins: Ratios of margins to add around the data inside
-                            the plot area for each side
-        :type dataMargins: A 4-tuple of float as (xMin, xMax, yMin, yMax).
-        """
-        pass
-
     def setLimits(self, xmin, xmax, ymin, ymax, y2min=None, y2max=None):
         """Set the limits of the X and Y axes at once.
